+++ tw-tkms-starter/src/main/java/com/transferwise/kafka/tkms/ITkmsPaceMaker.java
@@ -7,7 +7,7 @@ public interface ITkmsPaceMaker {
 
   void doSmallPause(int shard);
 
-  Duration getLongWaitTime(int shard);
+  Duration getProxyStopTimeout(TkmsShardPartition tkmsShardPartition);
 
   Duration getPollingPause(TkmsShardPartition tkmsShardPartition, int pollingBatchSize, int polledMessagesCount);