+++ tw-tkms-starter/src/main/java/com/transferwise/kafka/tkms/config/TkmsProperties.java
@@ -155,7 +155,7 @@ public class TkmsProperties implements InitializingBean {
    */
   @NotNull
   @jakarta.validation.constraints.NotNull
-  private Duration proxyTimeToLive = Duration.ofMinutes(10);
+  private Duration proxyTimeToLive = Duration.ofHours(1);
 
   /**
    * Safety net for validating message sizes before registering them with tw-tkms.
@@ -172,6 +172,13 @@ public class TkmsProperties implements InitializingBean {
    * <p>Can be useful in development environments, where environment owner wants to restrict resource usages globally.
    */
   private Duration minPollingInterval;
+
+  /**
+   * How long to wait for the `TkmsStorageToKafkaProxy` to stop, before giving up and logging a timeout error.
+   */
+  @NotNull
+  @jakarta.validation.constraints.NotNull
+  private Duration proxyStopTimeout = Duration.ofSeconds(15);
 
   /**
    * List topics used by the lib.
@@ -201,6 +208,19 @@ public class TkmsProperties implements InitializingBean {
    * Every shard can override settings from default configuration.
    */
   private Map<Integer, ShardProperties> shards = new HashMap<>();
+
+  /**
+   * When enabled, the messages are not immediately written to the database, but collected into memory and written to database just before commit.
+   * 
+   * <p>Allows to reduce transactions latency in case multiple individual messages are registered over the course of that transaction.
+   * The latency is reduced by batch inserting all the collected messages in the pre-commit hook.
+   * 
+   * <p>The tradeoff is that higher application memory is required for transactions sending out huge number of large messages.
+   * However, in practical applications, large transactions should be avoided anyway.
+   * 
+   * <p>May default to true for Postgres in upcoming versions. Or even default to true in all situations.
+   */
+  private boolean deferMessageRegistrationUntilCommit = false;
 
   @Valid
   @jakarta.validation.Valid
@@ -260,7 +280,9 @@ public class TkmsProperties implements InitializingBean {
     private Duration pollingInterval;
     private Duration pauseTimeOnErrors;
     private Integer insertBatchSize;
+    private Duration proxyStopTimeout;
     private boolean compressionOverridden;
+    private Boolean deferRegisteredMessagesUntilCommit;
     @Valid
     @jakarta.validation.Valid
     private Compression compression = new Compression();
@@ -275,8 +297,16 @@ public class TkmsProperties implements InitializingBean {
     private Map<String, String> kafka = new HashMap<>();
   }
 
+  public Duration getProxyStopTimeout(int shard) {
+    var shardProperties = shards.get(shard);
+    if (shardProperties != null && shardProperties.getProxyStopTimeout() != null) {
+      return shardProperties.getProxyStopTimeout();
+    }
+    return proxyStopTimeout;
+  }
+
   public String getTableBaseName(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.getTableBaseName() != null) {
       return shardProperties.getTableBaseName();
     }
@@ -284,7 +314,7 @@ public class TkmsProperties implements InitializingBean {
   }
 
   public DatabaseDialect getDatabaseDialect(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.getDatabaseDialect() != null) {
       return shardProperties.getDatabaseDialect();
     }
@@ -292,7 +322,7 @@ public class TkmsProperties implements InitializingBean {
   }
 
   public int getPartitionsCount(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.getPartitionsCount() != null) {
       return shardProperties.getPartitionsCount();
     }
@@ -300,7 +330,7 @@ public class TkmsProperties implements InitializingBean {
   }
 
   public int getPollerBatchSize(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.getPollerBatchSize() != null) {
       return shardProperties.getPollerBatchSize();
     }
@@ -308,15 +338,23 @@ public class TkmsProperties implements InitializingBean {
   }
 
   public Duration getPollingInterval(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.getPollingInterval() != null) {
       return shardProperties.getPollingInterval();
     }
     return pollingInterval;
   }
 
+  public boolean deferMessageRegistrationUntilCommit(int shard) {
+    var shardProperties = shards.get(shard);
+    if (shardProperties != null && shardProperties.getDeferRegisteredMessagesUntilCommit() != null) {
+      return shardProperties.getDeferRegisteredMessagesUntilCommit();
+    }
+    return deferMessageRegistrationUntilCommit;
+  }
+
   public Duration getPauseTimeOnErrors(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.getPauseTimeOnErrors() != null) {
       return shardProperties.getPauseTimeOnErrors();
     }
@@ -324,7 +362,7 @@ public class TkmsProperties implements InitializingBean {
   }
 
   public int getInsertBatchSize(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.getInsertBatchSize() != null) {
       return shardProperties.getInsertBatchSize();
     }
@@ -332,7 +370,7 @@ public class TkmsProperties implements InitializingBean {
   }
 
   public Compression getCompression(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.isCompressionOverridden()) {
       return shardProperties.getCompression();
     }
@@ -340,7 +378,7 @@ public class TkmsProperties implements InitializingBean {
   }
 
   public EarliestVisibleMessages getEarliestVisibleMessages(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.getEarliestVisibleMessages() != null) {
       return shardProperties.getEarliestVisibleMessages();
     }
@@ -348,7 +386,7 @@ public class TkmsProperties implements InitializingBean {
   }
 
   public List<Integer> getDeleteBatchSizes(int shard) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.deleteBatchSizes != null && !shardProperties.deleteBatchSizes.isEmpty()) {
       return shardProperties.deleteBatchSizes;
     }
@@ -357,7 +395,7 @@ public class TkmsProperties implements InitializingBean {
   }
 
   public NotificationLevel getNotificationLevels(int shard, NotificationType type) {
-    ShardProperties shardProperties = shards.get(shard);
+    var shardProperties = shards.get(shard);
     if (shardProperties != null && shardProperties.notificationLevels.get(type) != null) {
       return shardProperties.notificationLevels.get(type);
     }
@@ -409,6 +447,20 @@ public class TkmsProperties implements InitializingBean {
     private String tableName = "tw_tkms_earliest_visible_messages";
 
     private Duration lookBackPeriod = Duration.ofMinutes(5);
+
+    /**
+     * Interval, after we will poll all records, even when earliest message system has calculated a look-back period.
+     *
+     * <p>This is for situations where an application have a risk of still having long-running transactions, registering messages over its duration.
+     *
+     * <p>Notice however, that the order of messages may change in those situations.
+     * 
+     * <p>In case this is set, the first poll without id limit will be done when `StorageToKafkaProxy` acquires a lock.
+     * I.e. when a new node starts, or the same nodes starts again, to proxy messages from the database to the Kafka.
+     * 
+     * <p>In most cases you want to set this higher than `proxyTimeToLive`, so it will happen only once per proxy runtime.
+     */
+    private Duration pollAllInterval = null;
   }
 
   @Data