+++ tw-tkms-starter/src/test/java/com/transferwise/kafka/tkms/PostgresEndToEndIntTest.java
@@ -1,5 +1,7 @@
 package com.transferwise.kafka.tkms;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -10,20 +12,28 @@ import org.springframework.test.context.ActiveProfiles;
 public class PostgresEndToEndIntTest extends EndToEndIntTest {
 
   private static Stream<Arguments> compressionInput() {
-    return Stream.of(
-        Arguments.of(CompressionAlgorithm.GZIP, 111, 110),
-        Arguments.of(CompressionAlgorithm.NONE, 1171, 1171),
-        Arguments.of(CompressionAlgorithm.LZ4, 134, 134),
-        Arguments.of(CompressionAlgorithm.SNAPPY, 160, 160),
-        Arguments.of(CompressionAlgorithm.SNAPPY_FRAMED, 158, 158),
-        Arguments.of(CompressionAlgorithm.ZSTD, 100, 100)
-    );
+    var deferUntilCommits = List.of(false, true);
+    var arguments = new ArrayList<Arguments>();
+
+    for (var deferUntilCommit : deferUntilCommits) {
+      return Stream.of(
+          Arguments.of(CompressionAlgorithm.GZIP, 111, 110, deferUntilCommit),
+          Arguments.of(CompressionAlgorithm.NONE, 1171, 1171, deferUntilCommit),
+          Arguments.of(CompressionAlgorithm.LZ4, 134, 134, deferUntilCommit),
+          Arguments.of(CompressionAlgorithm.SNAPPY, 160, 160, deferUntilCommit),
+          Arguments.of(CompressionAlgorithm.SNAPPY_FRAMED, 158, 158, deferUntilCommit),
+          Arguments.of(CompressionAlgorithm.ZSTD, 100, 100, deferUntilCommit)
+      );
+    }
+
+    return arguments.stream();
   }
 
   @ParameterizedTest
   @MethodSource("compressionInput")
   @Override
-  public void testMessageIsCompressed(CompressionAlgorithm algorithm, int expectedSerializedSize, int expectedSerializedSizeAlt) {
-    super.testMessageIsCompressed(algorithm, expectedSerializedSize, expectedSerializedSizeAlt);
+  public void testMessageIsCompressed(CompressionAlgorithm algorithm, int expectedSerializedSize, int expectedSerializedSizeAlt,
+      boolean deferUntilCommit) {
+    super.testMessageIsCompressed(algorithm, expectedSerializedSize, expectedSerializedSizeAlt, deferUntilCommit);
   }
 }